refactoring(backoffice) : sépare les contenu du backoffice en différent composant pour plus de maintenabilité
