+++ frontend/src/pages/BackofficeAdmin.tsx
@@ -1,239 +1,12 @@
 // style:
-import React, { useState, type FormEvent } from "react"
 import "../scss/pages/backoffice.scss"
-import { MapContainer, TileLayer, Marker, Popup } from 'react-leaflet';
-import L from 'leaflet';
-import { useCreateCategoryMutation, useGetAllCategoriesQuery } from "../generated/graphql-types";
+import React, { useState, type FormEvent } from "react"
 
-import useStyleColors from "./backofficeHandler/styleColors";
-import useImageVerificationFrontend from "./backofficeHandler/imageVerification";
-
-type Category = {
-	categoryName: string,
-	categoryId: number, 
-	style: string
-}
+import BackofficeCategory from "../components/BackofficeCategory";
+import BackofficeCity from "../components/BackofficeCity";
+import BackofficeUser from "../components/BackofficeUser";
 
 export default function BackofficeAdmin() {
-	
-	// handle the admin tab + get and set the date
-	const [adminTabCities, setAdminTabCities] = useState('createCity')
-	const handleAdminTabCities = (tab: string) => {
-		setAdminTabCities(tab);
-	}
-
-	// to handle coordinateFormatError, see below
-	const [mapLatitude, setMapLatitude] = useState(0);
-	const [isLatitudeValid, setIsLatitudeValid] = useState(true)
-	const [mapLongitude, setMapLongitude] = useState(0);
-	const [isLongitudeValid, setIsLongitudeValid] = useState(true);
-	const [showMap, setShowMap] = useState(false);
-	let coordinateFormatError = ""
-	const checkCoordinateInput = (e: React.ChangeEvent<HTMLInputElement>, setValid: (valid: boolean) => void, type: string) => {
-		const coordinate = Number(e.target.value);
-		console.log('checkCoordinate : ', e, type, coordinate)
-
-		// // 1. Decimal Degrees (DD): 48.8566, -123.3847
-		// const regex = type === "latitude"
-		// 	? /^-?([0-8]?\d(\.\d+)?|90(\.0+)?)$/
-		// 	: /^-?(1[0-7]\d(\.\d+)?|180(\.0+)?|\d{1,2}(\.\d+)?)$/;
-		// // Examples: 48.8566, -123.3847, 90, -180
-		// const match = coordinate.match(regex)
-		console.log('match in check coordinate : ', {
-			// regex: regex,
-			coordinate: coordinate,
-			type: type,
-			// try: (coordinate.match(regex)),
-			// match: match
-		})
-		if (coordinate) {
-			if (type === "latitude") {
-				if (coordinate <= 90 || coordinate >= -90) {
-					setMapLatitude(coordinate)
-					setIsLatitudeValid(true)
-				} else {
-					setIsLatitudeValid(false)
-					coordinateFormatError('Erreur de format. Veuillez choisir une latitude comprise entre 90 et -90 et une longitute comprise en 180 et -180')
-				}
-			}
-			if (type === "longitude") {
-				if (coordinate <= 180 || coordinate >= -180) {
-					 setIsLongitudeValid(false)
-				} else {
-					setIsLongitudeValid(false)
-				}
-			}
-			coordinateFormatError = "Veuillez renseigner une coordonnée."
-			return setValid(false)
-		}
-
-		// if (match) {
-		// 	console.log('there is a match in coordinate !')
-		// 	if (type === "latitude") {
-		// 		setMapLatitude(coordinate)
-		// 		setIsLatitudeValid(true)
-		// 	}
-		// 	if (type === "longitude") {
-		// 		setMapLongitude(coordinate)
-		// 		setIsLongitudeValid(true)
-		// 	}
-		// 	setValid(true)
-		// 	showMapHandler();
-		// } else {
-		// 	console.log('regex is false for coordinate')
-		// 	setIsLatitudeValid(false)
-		// 	setIsLongitudeValid(false)
-		// 	setValid(false)
-		// 	coordinateFormatError = "Le format de votre coordonnée n'est pas correcte. Veullez utilisez le format approprié : Degrées décimal. Par exemple : Longitude -48.876667 ; latitude : -123.393333 "
-		// }
-		return setValid
-	}
-
-	function showMapHandler() {
-		if (mapLatitude !== '' && mapLongitude !== '') {
-			setShowMap(true)
-		} else {
-			setShowMap(false)
-		}
-
-		console.log({
-			mapLatitude: mapLatitude,
-			mapLongitude: mapLongitude,
-		})
-		console.log('showMapHandler', showMap)
-	}
-
-
-	// to handle adding an image see below
-	const { checkFileSignature, verifyImageLoad } = useImageVerificationFrontend()
-	const [isImageValid, setImageValid] = useState('unverified');
-	const [imageError, setImageError] = useState('');
-	const validateImage = (e: React.ChangeEvent<HTMLInputElement>) => {
-		const file = e.target.files?.[0] || null;
-		console.log('in validImage ! file is : ', file)
-		setImageError('');
-		setImageValid('unverified');
-
-
-		if (!file) {
-			console.log('no file found ')
-			setImageError('Erreur lors du chargement du fichier. Veuillez réessayer.');
-			setImageValid('false');
-		} else {
-
-			const validType = ['image/jpeg', 'image/jpg', 'image/png', 'image/webp'];
-			console.log('file type is : ', file.type)
-			if (!validType.includes(file.type) || file === null) {
-				console.log('file type is not valid')
-				setImageError('Format invalide. Utilisez JPG, JPEG, PNG ou WEBP.')
-				setImageValid('false')
-				return e.target.value = '';
-			}
-
-			const maxSize = 5 * 1024 * 1024 // 5 MB in bytes
-			console.log('file size is : ', file.size, ' max size is : ', maxSize)
-			if (file.size > maxSize) {
-				console.log('file size is not valid')
-				setImageError('Image trop volumineuse. Taille Maximum : 5MB.');
-				setImageValid('false')
-				return e.target.value = '';
-			}
-			const imageSecurityCheck = async (file: File) => {
-				const fileMagicSignatureIsValid = await checkFileSignature(file);
-				console.log('in image security check, value of fileMagicSignatureIsValid = ', await checkFileSignature(file))
-				const fileVerifiedOnBrowserIsValid = await verifyImageLoad(file);
-				console.log('fileverifiedonbrowerisvalid ', fileVerifiedOnBrowserIsValid)
-				if (fileMagicSignatureIsValid === true && fileVerifiedOnBrowserIsValid === true) {
-						return setImageValid('true')
-				} else {
-					setImageError('Image invalide. Veuillez fournir une autre image (formats autorisés : png, jpeg, webp, jpg)');
-					setImageValid('false')
-					return e.target.value = '';
-				}
-			}
-			imageSecurityCheck(file)
-		}
-	}
-
-	const handleAddCitySubmit = async (e: FormEvent<HTMLFormElement>) => {
-		e.preventDefault();
-		const form = e.target;
-		const formAddCityData = new FormData(form as HTMLFormElement);
-		const formJsonAddCity = Object.fromEntries(formAddCityData.entries())
-	}
-
-	const [adminTabCategories, setAdminTabsCategories] = useState('add-categories');
-	const { data: allCategoriesData, loading: allCategoriesLoading, error: allCategoriesError } = useGetAllCategoriesQuery();
-	const [importedColors, setImportedColors] = useState(true); 
-	const { colors } = useStyleColors()
-	const importedColorsHandler = () => {
-		if (!colors) {
-			return setImportedColors(false)
-		} else {
-			return setImportedColors(true)
-		}
-	}
-	const handleAdminTabCategories = (tab: string) => {
-		setAdminTabsCategories(tab); 
-		if (adminTabCategories === 'add-categories') importedColorsHandler()
-	}
-
-	const [createCategory] = useCreateCategoryMutation()
-	const handleAddCategory = async (e: FormEvent<HTMLFormElement>) => {
-		console.log("handleAddCategory has been called ! ")
-		e.preventDefault();
-		const form = e.target;
-		console.log("form is : ", form)
-		const formAddCategoryData = new FormData(form as HTMLFormElement);
-		console.log("formAddCategoryData is : ", formAddCategoryData)
-		const fromJsonAddCategory = Object.fromEntries(formAddCategoryData.entries())
-		console.log("formJsonAddCategory : ", fromJsonAddCategory)
-
-		try {
-			const result = await createCategory({
-				variables: {
-					data: {
-						categoryName: fromJsonAddCategory['categoryName'] as string,
-						style: fromJsonAddCategory['style'] as string
-					}
-				}
-			});
-
-			if (result.data) {
-				alert('Category created ! ')
-				console.log('category : ', result.data)
-			}
-			if (result.errors) {
-				throw new Error(result.errors[0].message);
-			}
-		} catch (error) {
-			console.error('Error creating new category : ', error)
-		}
-	}
-	const [editCategory, setEditCategory] = useState(false); 
-	const [editCategoryName, setEditCategoryName] = useState('')
-	const [editCategoryStyle, setEditCategoryStyle] = useState('')
-	const handleEditCategory = (name:string, style:string) => {
-		if (editCategory === true) setEditCategory(false)
-		if (editCategory === false) setEditCategory(true)
-		
-			if (editCategoryName === name){
-				setEditCategory(false)
-			} else {
-				setEditCategory(true)
-				setEditCategoryName(name);
-				setEditCategoryStyle(style);
-			}
-			//TODO : add resolver here once it's done
-	}
-	const [userWantsToDeleteCategory, setUserWantsToDeleteCategory] = useState(false); 
-	const handleDeleteCategory = () => {
-		// TODO : add resolver here once it's done
-		console.log('handle delete category has been called !')
-		setUserWantsToDeleteCategory(true)
-		return alert('Attention, supprimer la catégorie peut supprimer les POI qui y sont associé. Assurez-vous que cette catégorie n\'est associé à aucun POI')
-	}
-
 
 	const [isValidUser, setIsValidUser] = useState(false);
 	let validatedUserName = 'Urban';
@@ -245,183 +18,15 @@ export default function BackofficeAdmin() {
 					<h1>Administration</h1>
 				</div>
 				<section id="admin-ville">
-					<h2>Villes</h2>
-					<h3><div style={{ display: "inline-flex", flexDirection: "row", justifyContent: "flex-start", gap: "2rem", width: "100%" }}><div className={"tabBtn " + (adminTabCities === 'createCity' ? 'active' : '')} onClick={() => handleAdminTabCities('createCity')}>Ajouter une ville</div><div className={"tabBtn " + (adminTabCities === 'admin-city' ? 'active' : '')} onClick={() => handleAdminTabCities('admin-city')}>Administrer une ville</div></div></h3>
-
-					{/* ajouter une ville */}
-					{adminTabCities === "createCity" &&
-						<div className="input-field section-part">
-							<form onSubmit={handleAddCitySubmit}>
-								<label htmlFor="cityName">Nom de la ville
-									<input type="text" name="cityName" placeholder="Lyon" required />
-								</label>
-								<label htmlFor="description">Description
-									<input type="text" name="description" placeholder="Une ville de talent..." />
-								</label>
-								<label htmlFor="imageUrl">Image
-									<input type="file" name="imageUrl" placeholder="Image" accept="image/jpeg, image/png, image/jpg, image/webp" onChange={validateImage} />
-								{isImageValid === 'false' &&
-									<div className="img-div"><span className="img-div__error">{imageError}</span></div>
-								}
-								{isImageValid === 'true' &&
-									<div className="img-div">
-										<svg
-											xmlns="http://www.w3.org/2000/svg"
-											width="24"
-											height="24"
-											viewBox="0 0 24 24"
-										>
-											<path
-												d="M5 13 L9 17 L19 7"
-												fill="none"
-												stroke="#22c55e"
-												strokeWidth="2.5"
-												strokeLinecap="round"
-												strokeLinejoin="round"
-											/>
-										</svg>
-
-										<span className="img-div__valid">Image valide</span></div>}</label>
-								<p>Coordonnées</p>
-								<div className="add-ville-coordonnees">
-									<label htmlFor="cityLatitude">Latitude
-										<input type="number" name="cityLatitude" min="-90" max="90" placeholder="-48.876667" onBlur={(e) => checkCoordinateInput(e, setIsLatitudeValid, 'latitude')} required />
-										{isLatitudeValid === false && <p>{coordinateFormatError}</p>}
-									</label>
-									<label htmlFor="cityLongitude">Longitude
-										<input type="number" name="cityLongitude" min="-180" max="180" placeholder="-123.393333" onBlur={(e) => checkCoordinateInput(e, setIsLongitudeValid, 'longitude')} required />
-										{isLongitudeValid === false && <p>{coordinateFormatError}</p>}
-									</label>
-								</div>
-								{showMap &&
-									<MapContainer
-										center={[mapLatitude, mapLongitude]} // Paris
-										zoom={13}
-										style={{ height: '300px', width: '100%' }}
-									>
-										<TileLayer
-											url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-											attribution="&copy; OpenStreetMap contributors"
-										/>
-										<Marker position={[mapLatitude, mapLongitude]}>
-											<Popup>
-												Salut ! Je suis un marqueur Leaflet dans React.
-											</Popup>
-										</Marker>
-									</MapContainer>
-								}
-								{/* <input type="submit">Ajouter la ville</input> */}
-							</form>
-						</div>}
-
-					{/* administrer une ville */}
-					{adminTabCities === "admin-city" &&
-						<div className="admin-cities section-part">
-							<h3>Administrer une ville</h3>
-							<div className="admin-cities-wrapper">
-								<label htmlFor="select-ville">Choisissez une ville
-									<select name="select-ville">
-										<option>Saint-Loin-Lez-Allouètes</option>
-										<option>Machine Glin-glin</option>
-										<option>Lyon</option>
-									</select>
-								</label>
-							</div>
-						</div>}
+					<BackofficeCity />
 				</section>
 
-				{/* section utilisateur */}
-				<div className="separator"></div>
 				<section id="admin-users section-part">
-					<h2>Utilisateurs</h2>
-					<form className="section-part">
-						<label htmlFor="nom-utilisateur">Chercher le nom de l'utilisateur
-							<input type="text" name="nom-utilisateur" placeholder="Jean-Miche" required />
-						</label>
-						<input type="submit" name="isNameValid" value="Vérifier utilisateur" />
-					</form>
-					{!isValidUser &&
-						<div className="section-part">
-							<h3>Utilisateur '{validatedUserName}' trouvé</h3>
-							<ul>Rôle de cet utilisateur :
-								<li>Utilisateur standard</li>
-								<li>Créateur de point d'intérêt</li>
-							</ul>
-							<p>Voulez-vous ...</p>
-							<button>Administrer un rôle</button>
-							<button>Désactivé</button>
-							<button>Supprimer l'utilisateur</button>
-
-						</div>}
+					<BackofficeUser />
 				</section>
 
 				<section id="admin-categories">
-					<h2>Catégories</h2>
-					<h3><div style={{ display: "inline-flex", flexDirection: "row", justifyContent: "flex-start", gap: "2rem", width: "100%" }}><div className={"tabBtn " + (adminTabCategories === 'add-categories' ? 'active' : '')} onClick={() => handleAdminTabCategories('add-categories')}>Ajouter une catégorie</div><div className={"tabBtn " + (adminTabCategories === 'admin-categories' ? 'active' : '')} onClick={() => handleAdminTabCategories('admin-categories')}>Administrer les catégories</div></div></h3>
-					{adminTabCategories === "add-categories" &&
-						<div className="add-categories section-part">
-							<form onSubmit={handleAddCategory} className="flex-column">
-								<label htmlFor="categoryName">Nom de la catégorie à ajouter
-									<input type="text" name="categoryName" placeholder="Musée, restaurant..." required />
-									</label>
-									{importedColors &&
-									<label htmlFor="style">Sélectionnez une couleur de pin
-										<select name="style">
-											{Object.entries(colors).map(([key, value]) => (
-												<option key={key} value={key} style={{color:value}}>{key}</option>
-											))}
-										</select>
-									</label>
-									}
-								<input type="submit" value="Créer la nouvelle catégorie" />
-							</form>
-						</div>}
-					{adminTabCategories === 'admin-categories' &&
-						<div className="admin-categories section-part">
-							<label>Sélectionnez une catégorie</label>
-							<div className="flex-center">
-								{allCategoriesData &&
-									allCategoriesData?.getAllCategories.map((category: Category) => (
-										<div className="category-tag" key={category.categoryId} onClick={() => handleEditCategory(category.categoryName, category.style)}>
-											<div className="category-tag__pin" style={{backgroundColor:category.style}} ></div>
-											<div className="category-tag__name">{category.categoryName}</div>
-										</div>
-									))}
-									
-							</div>
-							{editCategory === true && 
-									<div className="flex-center input-field">Editer la catégorie
-										<label htmlFor="current-category-name">Nom actuel de la catégorie
-											<input type="text" name="current-category-name" value={editCategoryName} readOnly />
-										</label>
-										<label htmlFor="categoryName">Nouveau nom
-											<input type="text" name="categorName" placeholder={editCategoryName}/>
-										</label>
-										<label htmlFor="current-style">Couleur actuelle de la catégorie
-											<div style={{width:'50px', height:'50px', borderRadius:'90px', backgroundColor:editCategoryStyle}}></div>
-										</label>
-										<label htmlFor="style">Sélectionnez une nouvelle couleure de pin
-										<select name="style">
-											{Object.entries(colors).map(([key, value]) => (
-												<option key={key} value={key} style={{color:value}}>{key}</option>
-											))}
-										</select>
-									</label>
-									<input type="submit" value="modifier la catégorie"/>
-									<hr></hr>
-									<div className="danger-zone">
-									<span style={{textTransform:'uppercase'}}>Zone de danger !</span>
-									 {userWantsToDeleteCategory === false &&<button onClick={() => handleDeleteCategory()}>Supprimer la catégorie</button>}
-									 {userWantsToDeleteCategory === true && 
-									 <>
-										<button onClick={() => setUserWantsToDeleteCategory(false)}>Annuler la suppresion</button>
-										<input type="submit" value={`Confirmer la suppression de la catégorie '${editCategoryName}'`}/>
-									</>
-									 }
-									</div>
-									</div>
-							}
-						</div>}
+					<BackofficeCategory />
 				</section>
 			</div>
 		</>
